linter

+++ apps/test.c
@@ -1,7 +1,7 @@
 #include "testing.h"
 
 int main() {
-  for (int i = 0; i < ntests; ++i) {
+  for (u64 i = 0; i < ntests; ++i) {
     tests[i]();
   }
   return test_ret;
