+++ src/testing.c
@@ -1,5 +1,5 @@
 #include "testing.h"
 
-usize ntests = 0;
+u64 ntests = 0;
 test_func tests[2048];
 int test_ret = 0;
